boardv2 image fixes
Attempted to fix problems with images not showing in html.

"Restart Button.png" was renamed to "Restart_Button.png" to remove spaces in the file name and keep consistent with how the other image files were named.

"Alt" tags were added to the three images in boardv2.html, and a title tag was added to the head. Slight adjustments to indentation were also made within the body tags.

The code lines for the three images are lines 84, 107, and 142.